refactor: CLI, WEB용 폴더 묶음, 공통적으로 쓰이는건 common으로
