+++ index.html
@@ -4,12 +4,12 @@
     <meta charset="UTF-8" />
     <title>🎱 행운의 로또</title>
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <link rel="stylesheet" href="./src/css/reset.css" />
-    <link rel="stylesheet" href="./src/css/style.css" />
-    <link rel="stylesheet" href="./src/css/lotto.css" />
-    <link rel="stylesheet" href="./src/css/modal.css" />
-    <link rel="stylesheet" href="./src/css/header.css" />
-    <link rel="stylesheet" href="./src/css/footer.css" />
+    <link rel="stylesheet" href="./src/web/css/reset.css" />
+    <link rel="stylesheet" href="./src/web/css/style.css" />
+    <link rel="stylesheet" href="./src/web/css/lotto.css" />
+    <link rel="stylesheet" href="./src/web/css/modal.css" />
+    <link rel="stylesheet" href="./src/web/css/header.css" />
+    <link rel="stylesheet" href="./src/web/css/footer.css" />
   </head>
 
   <body>
