+++ src/common/domain/LottoGame.js
