+++ src/common/domain/LottoMaker.js
