+++ src/web/css/lotto.css
