+++ src/web/view/renderLottoList.js
